# Conflicts:
#	src/main/java/dev/rdh/argentum/mixin/core/render/FrustumCullerMixin.java

+++ src/main/java/dev/rdh/argentum/impl/render/instancing/InstancedGeometryBuffer.java
@@ -12,6 +12,7 @@ import org.lwjgl.opengl.GL11;
 import org.lwjgl.system.MemoryUtil;
 
 import java.nio.FloatBuffer;
+import java.nio.IntBuffer;
 
 public final class InstancedGeometryBuffer {
     private FloatBuffer vertices;
@@ -40,8 +41,7 @@ public final class InstancedGeometryBuffer {
                     TessellationBinding.forVertexBuffer(this.vertexBuffer, this.vertexFormat),
                     TessellationBinding.forVertexBuffer(this.instanceBuffer, this.instanceFormat,
                             this.vertexFormat.getAttributes().size(), 1)
-            }
-            );
+            });
             this.tessellation.init(commandList);
             this.vertices = null;
         } catch (RuntimeException exception) {
@@ -50,9 +50,9 @@ public final class InstancedGeometryBuffer {
         }
     }
 
-    public void draw(CommandList commandList, FloatBuffer instances, int vertexCount, int instanceCount) {
+    public void draw(CommandList commandList, IntBuffer instances, int vertexCount, int instanceCount) {
         this.initialize(commandList);
-        commandList.uploadData(this.instanceBuffer, MemoryUtil.memAddress(instances), (long)instances.remaining() * Float.BYTES, GlBufferUsage.STREAM_DRAW);
+        commandList.uploadData(this.instanceBuffer, MemoryUtil.memAddress(instances), (long)instances.remaining() * Integer.BYTES, GlBufferUsage.STREAM_DRAW);
         this.tessellation.bind(commandList);
         try {
             ARBDrawInstanced.glDrawArraysInstancedARB(GL11.GL_QUADS, 0, vertexCount, instanceCount);
